move satellite.py contents into forest.drivers.eida50
